Two bugfixes and a cleanup 1. When loading boolean settings we have to explicitly check for the return of nil from the self.settings:readSetting() function and only modify the default if the return value is NOT nil.
2. When closing the document we have to reset the settings to defaults,
otherwise their values are re-used on the next open.

3. The messages "Links on page ON/OFF" changed to
"Link underlines ON/OFF" because the old version gave two false
impressions: a) as if the action is pertinent only to the current page
(but in fact it is global for the entire document) and b) that we are
disabling the links whereas we are only disabling the show of underlines
under the links. The new message addresses both issues.