Use file:lines() instead of io.lines(filename) API This is due to a very strange quirk in Lua. If we use io.lines(filename) API like this:
for line in io.lines(filename) do
  ....
  if something then break end
end

Then the file descriptor table is growing infinitely because the
descriptor is not closed on breaking out of io.lines() loop. So we have
to use the ordinary file api like this:

local file = io.open(filename, "r")
for line in file:lines() do
  ....
  if something then break end
end
file:close()

This way there is no file descriptor leak.