bugfix: readerchooser dialog should not be visible when opening a file
when select a reader to open a document the readerchooser dialog is
overlayed with opening file notification window. The overlayed three
windows(file chooser, reader chooser and notification window) will make
users feel unconfortable easpecially when opening in koptreader which
usually freezes the screen for several seconds.

With this patch the reader chooser dialog is closed before the opening
file noticifation is shown up.