Adds zip ext to PDFReader and makes extensions accesible through defaults.lua
A lot of manga comes as .zip, and not .cbz (for all intents and purposes, it
is the same thing, just the extension is different). This allows the user to
choose the reader to open her zip files - PDFReader or CREReader, in the same
way she can choose to use PDFReader or PDFReflow for pdf files. I also moved
the list of supported extensions to defaults.lua.