bugfix: FileChooser crashes when closing landscape document
Yes, we already fixed that forcing portrait mode on KPV start, and
document close. However, this is a fix that will prevent the crash if
we don't force portrait. I'm not suggesting that we stop forcing portrait -
I think we all agreed that FileChooser in landscape mode on K3 is not
very usable. But, this might be handy on Touch or KPW (we'll see about that
later), so it should be usable should we need it. To test this, you need to
comment out self:setRotationMode(0) line in UniReader:inputLoop() in
unireader.lua, thus removing forcing of portrait on document close
.