fix for Kobo framebuffers: allow for rotation at framebuffer setup
the Kobo framebuffer comes up in landscape mode, that is, landscape
in the other direction as we do it when going into landscape mode.
we leave it this way, and rather allow for a rotation correction at
framebuffer setup.

this PR pulls the relevant changes from koreader-base and makes the
screen code aware of this additional rotation setting.

Kobo framebuffers should now work properly with the new framebuffer
FFI code. They can stay in 16bpp mode and don't change the current
framebuffer settings. They should thus properly interoperate with
the native framework on Kobos as well as sergeys KoboLauncher
(no garbled screen when returning)