Update readerscreenshot.lua
This adds screenshot functionality for the Kobo Aura. 

The pam-files generated can be read and converted using Imagemagick for linux/windows.

I've added the infomessage because it takes 5+ tries to get a screenshot by tapping the lower left and upper right corners. The other diagonal is not working for the Kobo Aura yet.