calculate maximum global cache size from free RAM In reflowing scroll mode with 2 pages hinting, 4 full page blitbuffers and koptcontexts should stay well in cache in the most demanding cases, with two pages shown on screen and two pages rendered in background. Since blitbuffer size is halved the size of page, we need cache size to be 6 times an average reflowed page size.
For Kobo Aura HD which has a resolution of 1440×1080, a reflowed page
could become 1080×4800. So 30MB of cache is demanded for this case.

This PR implements dynamic cache size allocating according to size of
system free memory. By default it will use 20 percent of free RAM with
a clip specified by DGLOBAL_CACHE_SIZE_MINIMUM and
DGLOBAL_CACHE_SIZE_MAXIMUM which are 10MB and 30MB respectively by default.