UPDATE readerfrontlight.lua
Code is added to switch between Kobo Aura (Phoenix) and others.

`ReaderFrontLight`, `self.ges_events` and the `function ReaderFrontLight:onAdjust` are now different for Kobo Aura. This enables me to do a two-fingered-pan over approx 1/4 Screenheight to change from full on to full off.

The difference to the default implementation is that both the `rate` and the `msg` are omitted and the `steps` are different.

FURTHER ROOM FOR IMPROVEMENT: It would be an improvement to get a final message after the two-fingered-pan is done stating the achieved setting.

Signed-off-by: Markismus <zulde.zuldemans@gmail.com>