Some more Kindle startup scrip tweaks.
First: sleep less. Looks okay on the slowest device I could find
(an indexing Touch in a hottish weather). Tweak if things are still racy
somewhere.

Second: Switch to a gentler approach to get rid of the title bar.
Seems to do the trick, and we don't get a slew of warnings in the logs
on exit ;).