Bring back the two KOBO variables for the light
This should close #856.
Although still closed I will check this one back with someone from the German forum, I hope this also closes #822, but there might be still some problems with the light-bug introduced in Kobo's firmware 3.5. But since 822 is already in the state closed we can leave things as they are for now and I will reopen 822 if necessary.