don't share info field in document We may have multiple documents opened simultaneously, such as in screensaver mode the current document is opened twice, it's better to keep seperate info table for each document.
This should fix #858. When a credocument is opened for screensaver to
get the cover page, the total pages info is set to 1(strangely enough),
before this fix this total pages is shared with the reading document
so the progress bar will get a progress percentage well beyond 100% and
rendering the progress bar out of the boundering box as shown in #858.