Update readerfooter.lua
In my opinion the logic was wrong.
IMagine to have a book with 20 <h3>s. Each of the <h3>2 has 10-20 <h4>s (all in all for sure more than 50), and some of these <h4> also have some <h5>s.
With the old logic you would see the <h5>s only since the <h4>s are too much.
But the <h3>s are way more important than the <h5>s, so the counter should start at the beginning.

I am unsure though if the for-loop should start with 1 or if we know somehow already that <h3> is the smallest?