Promote updates to full more agressively.
The goal is mainly to cover the directory browser.
Right now, this also catches navigating the menus, which is slightly
less desirable.
That said, popping a menu up and down *should* provoke a FULL flash,
but that's an improvement for another day, since we'd want to do it
without using a GC16 waveform mode... ;).