Bookmarks, CoverBrowser: scale dogear icon
The Dogear icon is 20x20 pixels and was never scaled where used. Now:

The bookmark icon (top right of screen) is scaled to 1/32th of the screen
width (previously, it was 1/30th on a 600px wide emulator, 1/53th on a GloHD).
On CreDocument, furthermore decrease its size if needed depending on the
selected margins so it never overwrite the text.

CoverBrowser list view: scale it to the available room under the
"N % of P page" text, so it does not cover "page".

CoverBrowser mosaic view: scale it to 1/16th of the cover rectangle, which
should prevent if from overwritting the text thanks to a max text width of
7/8 of the cover rectangle.

Also for CoverBrowser: don't index metadata for unsupported document
(which could happen when browsing files with PathChooser) and show
full filename for such documents.

Also: ImageWidget: small fix in case we use both scale_factor and
scale_for_dpi.