InputText: added isEditable() and isEdited()
Checks whether provided content can be given back
unaltered, which may not be the case after it is splitted
to UTF8 chars if the text is binary content.
Prevent editing text if that is the case.

Also accounts for the scrollbar width when measuring text
to prevent it from being displayed when not needed.

Also ensure a minimal size of the scrollbar thumb so it is
rendered when huge text with many lines is displayed.