Unbreak button flashing
Unfortunately, on most common buttons, we can't take any shortcuts :/.
Either we risk not seeing the flash, or, worse, getting tearing
artefacts because of the A2 updates :/.

So, yep, we have to go through two extra repaints (of the top-level
widget, because we can't repaint individual subwidgets),
and we have to delay the callback.

Yaaaay.