Fix BookStatusWidgets stars
Turned out to be broken when testing on a Touch.
Either there's a weird quirk on my Forma or my Touch,
or I've fixed spurious redraws since?

In any case, stars_container isn't a window, so the setDirty call wa
sbroken anyways.

Can't do targetet repaints, because we don't have the coordinates of the
stars container, AFAICT.

So just repaint the full widget.
No need for any delay, it should get merged with the Button unhilight w/
flash_ui, like everywhere else with that sort of interaction.