Always enforce a Portrait rotation via fbdepth
We were already doing it on the Forma, do it everywhere, so the 8bpp
blitting codepath has a chance to take the memcpy shortcuts...

Also, switch from MOD to AND, because MOD takes its sign from the
dividend in shell, which would be bad if our current computation didn't
ensure a positive unsigned value on that side of the divison.

(c.f., $(( (0 - 1) % 4 )) vs. $(( (0 - 1) & 3 )), which could
potentially be a viable computation for rotation, and which would
definitley be broken with a modulo).